requirements.txt
Flask==2.0.1
gunicorn==20.1.0
openai==1.3.0
python-dotenv==0.19.0
supabase==1.0.3
streamlit==1.24.0
flask-cors==4.0.0